Add support for network spaces
Automatically set the private-address relation key to the
network space binding for the relation the interface is
supporting for Juju >= 2.0; for older Juju versions continue
to use "unit-get private-address".

This change also adds the ability for the consuming charm
to override the hostname provided on the relation using the
option hostname parameter.

Change-Id: Iee0d1ac7168fe0c32deaf03add5cb261da6c2a79