Differentiate connected and ready hosts
rabbitmq_hosts is used to check which hosts are connected and also
which are ready. Split these into two different methods and update
the existing rabbitmq_hosts to be a wrapper around
rabbitmq_ready_hosts so that when charms build their contexts they
do not include rabbit units which are not ready for client
connections.

Change-Id: Ic8b669fedfef74ed24db733e01c5dfc0fd079abd
Closes-Bug: #1796886